requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
mysql_connector_repackaged==0.3.1
